pyproject.toml
[tool.poetry]
name = "rankings-django"
version = "0.1.0"
description = ""
authors = ["Simon Muller <samullers@gmail.com>"]
packages = [
    # Currently only used so we can get version with importlib.metadata.version
    { include = "rankings" }
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
Django = "3.2.20"
trueskill = "0.4.5"
Jinja2 = "2.11.3"
# Required by Jinja2 and v2.1.3 has "ImportError" of "soft_unicode"
markupsafe = "2.0.1"
django-mass-edit = "3.5.0"
djangorestframework = "3.14.0"
# Needed for DRF's dynamic schema generation for OpenAPI docs
# See: https://www.django-rest-framework.org/api-guide/schemas/#install-dependencies
uritemplate = "^4.1.1"

[tool.poetry.dev-dependencies]
flake8 = "^6.1.0"
black = "^23.7.0"
flake8-docstrings = "^1.6.0"
mypy = "^1.5.1"
coverage = "^7.3.0"

[tool.mypy]
# We only add configs here that can apply equally to our main code and our test code
ignore_missing_imports = true
exclude = ["migrations/"]  # "**/*/migrations/"

[tool.black]
exclude = "migrations/"

[tool.coverage.run]
branch = true
dynamic_context = "test_function"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
